Clean up project test factories
