+++ src/test/kotlin/helper/DeleteTaskFactory.kt
@@ -1,13 +1,11 @@
 package helpers
 
-import logic.entities.Project
+import helper.ProjectFactory.PROJECT1_ID
 import logic.entities.Task
 import java.util.UUID
 
 object DeleteTaskTestFactory {
 
-    val PROJECT1_ID: UUID = UUID.fromString("bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb")
-    val PROJECT2_ID: UUID = UUID.fromString("bbbbbbbb-cccc-bbbb-bbbb-bbbbbbbbbbbb")
 
     val TASK_1 = Task(
         id = UUID.fromString("11111111-1111-1111-1111-111111111111"),
@@ -23,26 +21,7 @@ object DeleteTaskTestFactory {
         state = "PENDING"
     )
 
-    val TASK_3 = Task(
-        id = UUID.fromString("33333333-3333-3333-3333-333333333333"),
-        name = "Deploy to production",
-        projectId = PROJECT2_ID,
-        state = "DONE"
-    )
 
-    val PROJECT_1 = Project(
-        id = UUID.randomUUID(),
-        name = "Gym App",
-        states = listOf("TODO", "IN_PROGRESS", "DONE"),
-        tasks = listOf(TASK_1, TASK_2)
-    )
 
-    val PROJECT_2 = Project(
-        id = PROJECT2_ID,
-        name = "Smart Home",
-        states = listOf("NEW", "ACTIVE", "ARCHIVED"),
-        tasks = listOf(TASK_3)
-    )
 
-    val ALL_PROJECTS = listOf(PROJECT_1, PROJECT_2)
 }
