+++ src/test/kotlin/logic/usecases/state/GetProjectStatesUseCaseTest.kt
@@ -7,9 +7,9 @@ import kotlinx.coroutines.test.runTest
 import logic.exception.PlanMateException.ValidationException.InvalidProjectIDException
 import logic.exception.PlanMateException.NotFoundException.ProjectNotFoundException
 import logic.repository.ProjectRepository
-import helper.GetProjectStatesUseCaseTestFactory.dummyProject
-import helper.GetProjectStatesUseCaseTestFactory.EXISTING_PROJECT_ID
-import helper.GetProjectStatesUseCaseTestFactory.INVALID_PROJECT_ID
+import helper.ProjectFactory.dummyProject
+import helper.ProjectFactory.EXISTING_PROJECT_ID
+import helper.ProjectFactory.INVALID_PROJECT_ID
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
