+++ src/test/kotlin/logic/usecases/state/GetTaskStateUseCaseTest.kt
@@ -7,11 +7,11 @@ import kotlinx.coroutines.test.runTest
 import logic.exception.PlanMateException.NotFoundException.TaskNotFoundException
 import logic.exception.PlanMateException.ValidationException.InvalidTaskIDException
 import logic.repository.TaskRepository
-import helper.GetTaskStateUseCaseTestFactory.EXPECTED_TASK_STATE
-import helper.GetTaskStateUseCaseTestFactory.existingTask
-import helper.GetTaskStateUseCaseTestFactory.existingTaskID
-import helper.GetTaskStateUseCaseTestFactory.invalidTaskID
-import helper.GetTaskStateUseCaseTestFactory.notExistingTaskID
+import helper.ProjectFactory.EXPECTED_TASK_STATE
+import helper.ProjectFactory.existingTask
+import helper.ProjectFactory.EXISTING_TASK_ID
+import helper.ProjectFactory.INVALID_TASK_ID
+import helper.ProjectFactory.notExistingTaskID
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
 import kotlin.test.Test
@@ -33,7 +33,7 @@ class GetTaskStateUseCaseTest {
         coEvery { repository.getTaskById(any()) } returns existingTask
 
         //When
-        val state = useCase.execute(existingTaskID)
+        val state = useCase.execute(EXISTING_TASK_ID)
 
         //Then
         assertThat(state).isEqualTo(EXPECTED_TASK_STATE)
@@ -42,7 +42,7 @@ class GetTaskStateUseCaseTest {
     @Test
     fun `should throw InvalidTaskIDException when the task ID is invalid`() = runTest{
         //Given & When & Then
-        assertThrows<InvalidTaskIDException> { useCase.execute(invalidTaskID) }
+        assertThrows<InvalidTaskIDException> { useCase.execute(INVALID_TASK_ID) }
     }
 
     @Test
