+++ src/test/kotlin/ui/project/CreateProjectUIControllerTest.kt
@@ -1,13 +1,13 @@
 package ui.project
 
 import com.google.common.truth.Truth.assertThat
+import helper.ProjectFactory.inValidProjectNameTest
+import helper.ProjectFactory.validProjectTest
 import io.mockk.*
 import kotlinx.coroutines.test.runTest
 import logic.exception.PlanMateException
 import logic.exception.PlanMateException.ValidationException.InvalidProjectNameException
 import logic.usecases.project.CreateProjectUseCase
-import logic.usecases.testFactories.CreateProjectTestFactory.inValidProjectNameTest
-import logic.usecases.testFactories.CreateProjectTestFactory.validProjectTest
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import ui.console.ConsoleIO
