+++ src/test/kotlin/ui/project/GetProjectUIControllerTest.kt
@@ -1,12 +1,12 @@
 package ui.project
 
 import com.google.common.truth.Truth.assertThat
+import helper.ProjectFactory.createProject
 import io.mockk.*
 import kotlinx.coroutines.test.runTest
 import logic.exception.PlanMateException.NotFoundException.ProjectNotFoundException
 import logic.exception.PlanMateException.ValidationException.InvalidUUIDFormatException
 import logic.usecases.project.GetProjectDetailsUseCase
-import helper.createProject
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
